+++ Source/minecraftplus_mod/CommonProxy.java
@@ -1,5 +1,6 @@
 package net.minecraftplus_mod;
 
+import cpw.mods.fml.common.Side;
 import net.minecraftforge.client.MinecraftForgeClient;
 
 public class CommonProxy{