+++ CHANGELOG.md
@@ -1,12 +1,23 @@
 Changelog
 =============
+r2b7
+-------------
+- Lots o' bug fixes, including:    
+	- Fixed Shield block bug    
+	- Fixed assorted server & client crashes    
+	- And more    
+- Added new Creative Tab    
+In the Creative Inventory, click the Right arrow on the top to see the Minecraft+ tab. Everything from the mod is grouped here now.    
+    
 r2b6
 -------------
 - Added new debugging/ stats file    
 - Added new folder in .minecraft for mod-related documents    
 - Fixed some bugs    
+	- Fixed block-related bugs
+	- More
 - Added a changelog (this) :D    
-
+    
 r2b5
 -------------
 - Fixed server crash bug